Rename model_artifacts cache module
