+++ inference/core/cache/model_artifacts.py
